Backend/Shop/app/serializers.py
from rest_framework import serializers
from app.models import Category, Product

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'name']

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'category', 'brand','title', 'description', 'price', 'discount_price']