Backend/Shop/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-js-asset==2.0.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
